To use CACurrentMediaTime as recommended by a few sources we need to define ccTime as a double.

+++ GameHack/libs/cocos2d/ccTypes.h
@@ -368,7 +368,7 @@ typedef enum {
 
 //! delta time type
 //! if you want more resolution redefine it as a double
-typedef float ccTime;
-//typedef double ccTime;
+//typedef float ccTime;
+typedef double ccTime;
 
 typedef float ccMat4[16];
